Performance improvements.
- Try to improve table drawing. Try to reuse existing table item views
and don't create them all the time. However, (using the profiler) the
problem seems to be the setting or drawing of the text in the table
view. Maybe also the resulting re-calculation of the column width causes
the performance drop...
- I tried to use a Handler to trigger the redraw a bit later. I though
removing previous update request from the message queue would improve
stuff. However, I was not able to make it work probably. The marker
where smoothly dragable for a bit but then hang because the table was
busy drawing. So I removed that approach.
- The poor man's solution I'm using now is to update the marker data
only on touch up. This makes it possible to drag all marker very
smoothly but data is not updated on the fly.