Make script fragments robust against screen rotations.
When rotating the screen fragments are managed automatically by the
system, i.e., they are restarted. For that reason they have to have a
default constructor. Previously the component was handed to the fragment
via the constructor. Since this does not work anymore the fragment
remembers the index of the component and gets it from activity on
restart.

Fix the ActivityStarterView. The assigned view id was not valid anymore,
after a rotation. Now the id is dependent on the order in which the
views are added to the fragment.