Use more descriptive names for experiment run and group
A run group is now called a run and a run is now a sensor. The
experiment structure is something like that:

Experiment:
- run0
-- sensor0
-- sensor1
- run1
-- sensor0
-- sensor1

The reason why I haven't choose these names before was that a sensor
doesn't need to be a single physical sensor but could be a combined
measurement of multiple sensors. However, that would be a multi-sensor
then... The new naming is much more intuitive.